storage/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
cryptography==41.0.4
distlib==0.3.7
Django==4.2.5
django-appconf==1.0.5
django-cors-headers==4.3.0
django-cryptography==1.1
django-elasticsearch-dsl==8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
elastic-transport==8.4.1
elasticsearch==8.10.0
elasticsearch-dsl==8.9.0
env==0.1.0
filelock==3.12.4
platformdirs==3.10.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.5
